requirements.txt
Flask==2.1.2
gunicorn==20.1.0
cycler==0.11.0
fonttools==4.33.3
kiwisolver==1.4.2
numpy==1.22.3
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0